user services and controllers

+++ src/main/java/com/supplyChain/users/user/UserRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.NoRepositoryBean;
 
 @NoRepositoryBean
-public interface UserRepository<T extends User> extends JpaRepository<T, Integer> {
+public interface UserRepository<T extends User> extends JpaRepository<T, Long> {
 }
